requirements.txt
Flask==3.0.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1
Werkzeug==3.0.3
requests==2.32.3
google-generativeai==0.8.2
